# models/UserModel.js
const mongoose = require('mongoose');

const userSchema = new mongoose.Schema({
  email: { type: String, required: true, unique: true },
  password: { type: String, required: true },
  name: { type: String, required: true },
  role: { type: String, required: true, enum: ['client', 'agent'] },
});

const User = mongoose.model('User', userSchema);

module.exports = User;